requirements.txt
playwright==1.42.0
Pillow==9.5.0
numpy==1.24.3
scipy==1.10.1
discord.py>=2.0.0
watchdog>=2.1.0
python-dotenv>=0.19.0
